requirements.txt
cycler==0.10.0
jieba==0.39
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
websocket-client==0.44.0
wordcloud==1.3.1
